refactor: move agent config to lib/agents/
