+++ src/lib/discovery/context-discovery.ts
@@ -5,7 +5,7 @@ import {
   parseInstructionContent,
   parsePromptContent,
 } from '../parsers/index.ts';
-import { targetAgents } from '../../target-agents.ts';
+import { targetAgents } from '../agents/index.ts';
 import type { ContextFormat, ContextType, DiscoveredItem, TargetAgent } from '../../types.ts';
 
 // ---------------------------------------------------------------------------
