+++ src/lib/install/context-check.ts
@@ -12,7 +12,7 @@ import {
   getLockEntriesByType,
 } from '../lock/index.ts';
 import { loadModelOverrides } from '../../model-aliases.ts';
-import { TARGET_AGENTS } from '../../target-agents.ts';
+import { TARGET_AGENTS } from '../agents/index.ts';
 import type { ContextType, LockEntry, TargetAgent } from '../../types.ts';
 
 // ---------------------------------------------------------------------------
