+++ src/lib/install/skill-installer.ts
@@ -15,7 +15,7 @@ import { join, basename, normalize, resolve, sep, relative, dirname } from 'path
 import { homedir, platform } from 'os';
 import type { Skill, AgentType, RemoteSkill } from '../../types.ts';
 import type { WellKnownSkill } from '../../providers/wellknown.ts';
-import { agents, detectInstalledAgents, isUniversalAgent } from '../../agents.ts';
+import { agents, detectInstalledAgents, isUniversalAgent } from '../agents/index.ts';
 import { AGENTS_DIR, SKILLS_SUBDIR } from '../../constants.ts';
 import { parseSkillMd } from '../discovery/index.ts';
