+++ src/lib/transpilers/agent-transpilers.test.ts
@@ -8,7 +8,7 @@ import {
   transpileAgent,
   transpileAgentForAllAgents,
 } from './agent-transpilers.ts';
-import { TARGET_AGENTS } from '../../target-agents.ts';
+import { TARGET_AGENTS } from '../agents/index.ts';
 import type { CanonicalAgent, DiscoveredItem, TargetAgent } from '../../types.ts';
 
 // ---------------------------------------------------------------------------
