+++ src/lib/transpilers/agent-transpilers.ts
@@ -7,7 +7,7 @@ import type {
   TranspiledOutput,
 } from '../../types.ts';
 import { parseAgentContent } from '../parsers/index.ts';
-import { getTargetAgentConfig } from '../../target-agents.ts';
+import { getTargetAgentConfig } from '../agents/index.ts';
 import { resolveModel, type ModelOverrides } from '../../model-aliases.ts';
 import { quoteYaml } from '../../utils.ts';
 import { mergeOverrides } from '../parsers/index.ts';
