+++ src/restore.ts
@@ -3,7 +3,7 @@ import pc from 'picocolors';
 import { readLocalLock } from './lib/lock/index.ts';
 import { runAdd } from './add.ts';
 import { runSync, parseSyncOptions } from './sync.ts';
-import { getUniversalAgents } from './agents.ts';
+import { getUniversalAgents } from './lib/agents/index.ts';
 import { readDotaiLock, getLockEntriesByType } from './lib/lock/index.ts';
 import { addPrompts, addAgents } from './lib/install/index.ts';
 import { parseSource } from './lib/parsers/index.ts';
