+++ tests/e2e-utils.ts
@@ -19,7 +19,7 @@ import { execSync } from 'child_process';
 import type { ContextType, LockEntry, TargetAgent } from '../src/types.ts';
 import type { DotaiLockFile } from '../src/lib/lock/index.ts';
 import { readDotaiLock } from '../src/lib/lock/index.ts';
-import { targetAgents } from '../src/target-agents.ts';
+import { targetAgents } from '../src/lib/agents/index.ts';
 
 // ---------------------------------------------------------------------------
 // Temp project directory
