+++ tests/list-installed.test.ts
@@ -3,7 +3,7 @@ import { mkdir, writeFile, rm } from 'fs/promises';
 import { join } from 'path';
 import { tmpdir } from 'os';
 import { listInstalledSkills } from '../src/lib/install/index.ts';
-import * as agentsModule from '../src/agents.ts';
+import * as agentsModule from '../src/lib/agents/agents.ts';
 
 describe('listInstalledSkills', () => {
   let testDir: string;
